requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
instagram-scraper==1.5.19
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.2
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.4.3
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
PyPrind==2.11.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scandir==1.6
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
tensorflow==1.4.0rc0
tensorflow-tensorboard==0.4.0rc1
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.1.0
